Charts
I am aware that on the notes Galaan was in charge of this portion right after I already worked on this. If it is okay with Galaan and the rest of you guys I would like to continue with this portion but if Galaan already did this part as well I will definitely start another one. :)